refactor : useNavigation hook을 통한 관심사 분리
