+++ src/shared/components/Navigation/hooks/useNavigation.ts
@@ -1,0 +1,25 @@
+import { useContext } from 'react';
+import { useLocation } from 'react-router-dom';
+import { NAV_CONFIG } from '@/constants/navigation';
+import { AuthContext } from '@/providers/auth';
+import { replaceRouteParams } from '@/utils/replaceRouteParams';
+import type { NavItemData } from '@/types/navigation';
+
+export const useNavigation = () => {
+  const { user, logout } = useContext(AuthContext);
+  const location = useLocation();
+  const currentRoute = location.pathname;
+
+  const items: NavItemData[] = NAV_CONFIG[user?.role ?? 'guest'];
+  const leftItems = items.filter((item) => !['login', 'logout'].includes(item.key));
+  const rightItem = items.find((item) => ['login', 'logout'].includes(item.key));
+
+  const getCurrentRoute = (item: NavItemData) => {
+    if (item.to?.includes(':clubId') && user?.clubId?.length) {
+      return replaceRouteParams(item.to, { clubId: user.clubId[0] });
+    }
+    return item.to || '#';
+  };
+
+  return { user, logout, leftItems, rightItem, getCurrentRoute, currentRoute };
+};
