+++ src/shared/components/Navigation/index.tsx
@@ -1,31 +1,11 @@
 import styled from '@emotion/styled';
-import { useContext } from 'react';
-import { useLocation } from 'react-router-dom';
-import { NAV_CONFIG } from '@/constants/navigation';
-import { AuthContext } from '@/providers/auth';
 import { replaceRouteParams } from '@/utils/replaceRouteParams';
+import { useNavigation } from './hooks/useNavigation';
 import { NavigationContainer } from './NavigationContainer';
 import { NavigationItem } from './NavigationItem';
-import type { NavItemData } from '@/types/navigation';
 
 export const Navigation = () => {
-  const location = useLocation();
-  const currentRoute = location.pathname;
-
-  const { user } = useContext(AuthContext);
-  const { logout } = useContext(AuthContext);
-
-  const items: NavItemData[] = NAV_CONFIG[user?.role ?? 'guest'];
-
-  const leftItems = items.filter((item) => !['login', 'logout'].includes(item.key));
-  const rightItem = items.find((item) => ['login', 'logout'].includes(item.key));
-
-  const getCurrentRoute = (item: NavItemData) => {
-    if (item.to?.includes(':clubId') && user?.clubId?.length) {
-      return replaceRouteParams(item.to, { clubId: user.clubId[0] });
-    }
-    return item.to || '#';
-  };
+  const { user, logout, leftItems, rightItem, getCurrentRoute, currentRoute } = useNavigation();
 
   const handleItemClick = (key: string) => {
     if (key == 'logout') {
